refactor: interface ServerResponse위치 변경 및 useQuery에 누락된 타입 선언
