+++ api/Group/index.ts
@@ -1,9 +1,9 @@
 import { GroupColor } from '@/constants';
+import { ServerResponse } from '@/types';
 import api from '..';
 
 export interface GroupInfo {
   title: string;
-  //   type: GroupType;
   type: string;
   coverColor: GroupColor;
 }
@@ -14,11 +14,6 @@ export interface GroupId {
 
 export interface GroupNickname {
   nickname: string;
-}
-
-export interface ServerResponse<T> {
-  message: string;
-  content: T;
 }
 
 export interface GroupDetail {
@@ -68,29 +63,29 @@ export const updateGroup = async (updateGroupInfo: GroupInfo & GroupId): Promise
   return data;
 };
 
-export const deleteGroup = async (groupId: string): Promise<ServerResponse<null>> => {
+export const deleteGroup = async (groupId: string): Promise<ServerResponse> => {
   const { data } = await api.delete(`/api/group/${groupId}`);
   return data;
 };
 
-export const joinGroup = async (info: GroupNickname & GroupId): Promise<ServerResponse<null>> => {
+export const joinGroup = async (info: GroupNickname & GroupId): Promise<ServerResponse> => {
   const { nickname, groupId } = info;
   const { data } = await api.post(`/api/group/${groupId}/participant`, { nickname });
   return data;
 };
 
-export const changeAdmin = async (info: GroupNickname & GroupId): Promise<ServerResponse<null>> => {
+export const changeAdmin = async (info: GroupNickname & GroupId): Promise<ServerResponse> => {
   const { nickname, groupId } = info;
   const { data } = await api.patch(`/api/group/admin/${groupId}`, { nickname });
   return data;
 };
 
-export const withdrawalGroup = async (groupId: string): Promise<ServerResponse<null>> => {
+export const withdrawalGroup = async (groupId: string): Promise<ServerResponse> => {
   const { data } = await api.delete(`/api/group/${groupId}`);
   return data;
 };
 
-export const changeNickname = async (info: GroupNickname & GroupId): Promise<ServerResponse<null>> => {
+export const changeNickname = async (info: GroupNickname & GroupId): Promise<ServerResponse> => {
   const { nickname, groupId } = info;
   const { data } = await api.patch(`/api/participant/${groupId}`, { nickname });
   return data;
