+++ queries/Group/index.ts
@@ -8,8 +8,7 @@ export { useChangeNickname } from './useChangeNickname';
 export { useChangeAdmin } from './useChangeAdmin';
 export { useDeleteGroup } from './useDeleteGroup';
 export { useWithdrawalGroup } from './useWithdrawalGroup';
-
-import { ServerResponse } from '@/api/Group';
+import { ServerResponse } from '@/types';
 
 export const message = {
   onSuccess: (data: ServerResponse<any>) => console.log(data.message),
