+++ queries/Group/useChangeAdmin.ts
@@ -1,8 +1,9 @@
 import { message } from './index';
-import { changeAdmin, GroupId, GroupNickname, ServerResponse } from '@/api/Group';
+import { changeAdmin, GroupId, GroupNickname } from '@/api/Group';
 import { useMutation } from '@tanstack/react-query';
 import { AxiosError } from 'axios';
+import { ServerResponse } from '@/types';
 
 export const useChangeAdmin = () => {
-  return useMutation<ServerResponse<null>, AxiosError, GroupNickname & GroupId>(changeAdmin, message);
+  return useMutation<ServerResponse, AxiosError, GroupNickname & GroupId>(changeAdmin, message);
 };
