+++ queries/Group/useGroupDetail.ts
@@ -1,6 +1,7 @@
-import { getGroupDetail } from '@/api/Group';
+import { getGroupDetail, GroupDetail } from '@/api/Group';
+import { ServerResponse } from '@/types';
 import { useQuery } from '@tanstack/react-query';
 
 export const useGroupDetail = (groupId: string) => {
-  return useQuery(['groupDetail', groupId], () => getGroupDetail(groupId));
+  return useQuery<ServerResponse<GroupDetail>>(['groupDetail', groupId], () => getGroupDetail(groupId));
 };
