+++ queries/Group/useJoinGroup.ts
@@ -1,8 +1,9 @@
 import { message } from './index';
-import { GroupId, GroupNickname, joinGroup, ServerResponse } from '@/api/Group';
+import { GroupId, GroupNickname, joinGroup } from '@/api/Group';
 import { useMutation } from '@tanstack/react-query';
 import { AxiosError } from 'axios';
+import { ServerResponse } from '@/types';
 
 export const useJoinGroup = () => {
-  return useMutation<ServerResponse<null>, AxiosError, GroupNickname & GroupId>(joinGroup, message);
+  return useMutation<ServerResponse, AxiosError, GroupNickname & GroupId>(joinGroup, message);
 };
