+++ queries/Group/useParticipantList.ts
@@ -1,6 +1,7 @@
-import { getParticipantList } from '@/api/Group';
+import { getParticipantList, ParticipantList } from '@/api/Group';
+import { ServerResponse } from '@/types';
 import { useQuery } from '@tanstack/react-query';
 
 export const useParticipantList = (groupId: string) => {
-  return useQuery(['participantList', groupId], () => getParticipantList(groupId));
+  return useQuery<ServerResponse<ParticipantList>>(['participantList', groupId], () => getParticipantList(groupId));
 };
